refactor(batch): extract RefreshReferenceMode into its own module
Moves the `RefreshReferenceMode` enum + its `impl` (with the
`is_trusted()` helper) out of the now-large `batch/mod.rs` into a
dedicated `batch/refresh_reference_mode.rs` module. The type is
re-exported from `batch` so existing call sites (e.g.
`crate::batch::RefreshReferenceMode`) continue to work without
changes.

No behavior change. The intra-doc links to `Element::Reference`,
`Element::ReferenceWithSumItem`, and `GroveOp::RefreshReference`
are spelled with explicit `crate::` / `super::` paths since the
type now lives in a sibling submodule.

1596 grovedb tests pass.
